Not catching the XMLIOException anymore as a fallback for the deprecated KblValidation
There is now a semi-breaking change.

Before, a validation exception and an IOException from reading a file
were both wrapped into a KblException.
An XMLIOException was not caught and the user had to handle it.

Now the IOException from reading the file is wrapped into an XMLIOException
means there are two sources for the XMLIOException:
1. The validation failed unexpectedly (was already the case).
2. The given file couldn't be read (is now the case).

The library will still not catch the XMLIOException and the user still has to handle it.

This means that the library will not throw an KblException anymore if reading the given file fails.
Instead an XMLIOException is thrown, similar to when validating a file errors out unexpectedly.

See https://github.com/4Soft-de/harness-model/pull/398#discussion_r2510054340
for the whole discussion.